+++ packages/gamut/src/Anchor/index.tsx
@@ -1,7 +1,7 @@
 import { styledOptions, system, variant } from '@codecademy/gamut-styles';
 import { StyleProps, variance } from '@codecademy/variance';
 import styled from '@emotion/styled';
-import { HTMLProps } from 'react';
+import { ComponentProps, forwardRef, HTMLProps, RefObject } from 'react';
 
 import { ButtonBase, ButtonSelectors } from '../ButtonBase/ButtonBase';
 
@@ -10,6 +10,23 @@ export interface AnchorProps
     StyleProps<typeof anchorVariants> {
   onClick?: HTMLProps<HTMLAnchorElement>['onClick'];
 }
+
+const outlineFocusVisible = {
+  [ButtonSelectors.OUTLINE]: {
+    content: "''",
+    position: 'absolute',
+    inset: -4,
+    borderRadius: '4px',
+    border: 2,
+    borderColor: 'primary',
+    opacity: 0,
+    zIndex: 0,
+  },
+
+  [ButtonSelectors.OUTLINE_FOCUS_VISIBLE]: {
+    opacity: 1,
+  },
+} as const;
 
 const anchorVariants = variant({
   base: {
@@ -22,16 +39,6 @@ const anchorVariants = variant({
     position: 'relative',
     color: 'primary',
     whiteSpace: 'nowrap',
-    [ButtonSelectors.OUTLINE]: {
-      content: "''",
-      position: 'absolute',
-      inset: -4,
-      borderRadius: '4px',
-      border: 2,
-      borderColor: 'primary',
-      opacity: 0,
-      zIndex: 0,
-    },
     [ButtonSelectors.HOVER]: {
       textDecoration: 'none',
       cursor: 'pointer',
@@ -41,29 +48,31 @@ const anchorVariants = variant({
       textDecoration: 'none',
       color: 'text-disabled',
     },
-    [ButtonSelectors.OUTLINE_FOCUS_VISIBLE]: {
-      opacity: 1,
-    },
   },
   variants: {
     standard: {
       color: 'primary',
+      fontWeight: 'bold',
+      WebkitFontSmoothing: 'antialiased',
+      MozOsxFontSmoothing: 'grayscale',
       [ButtonSelectors.HOVER]: {
         textDecoration: 'underline',
       },
       [ButtonSelectors.FOCUS_VISIBLE]: {
-        color: 'text',
+        WebkitFontSmoothing: 'antialiased',
+        MozOsxFontSmoothing: 'grayscale',
+        outline: 'none',
       },
+      ...outlineFocusVisible,
     },
     inline: {
       display: 'inline',
       whiteSpace: 'initial',
       textDecoration: 'underline',
-      [ButtonSelectors.OUTLINE]: {
-        display: 'none',
-      },
       [ButtonSelectors.FOCUS_VISIBLE]: {
-        outline: 'currentColor auto 4px',
+        outline: 'currentColor solid 2px',
+        borderRadius: '1px',
+        outlineOffset: '1.5px',
         textDecoration: 'underline',
       },
     },
@@ -75,7 +84,9 @@ const anchorVariants = variant({
       },
       [ButtonSelectors.FOCUS_VISIBLE]: {
         color: 'primary',
+        outline: 'none',
       },
+      ...outlineFocusVisible,
     },
   },
 });
@@ -91,8 +102,31 @@ export const AnchorBase = styled('a', styledOptions<'a'>())<AnchorProps>(
   anchorProps
 );
 
-export const Anchor = AnchorBase.withComponent(ButtonBase);
+type AnchorBaseProps =
+  | ComponentProps<typeof AnchorBase>
+  | (Exclude<ComponentProps<typeof AnchorBase>, 'ref'> &
+      ComponentProps<typeof ButtonBase>);
 
-Anchor.defaultProps = {
-  variant: 'inline',
-};
+export const Anchor = forwardRef<
+  HTMLAnchorElement | HTMLButtonElement,
+  AnchorBaseProps
+>(({ variant = 'inline', ...rest }, ref) => {
+  if (!rest.href) {
+    return (
+      <AnchorBase
+        as={ButtonBase}
+        variant={variant}
+        ref={ref as RefObject<HTMLAnchorElement>}
+        {...rest}
+      />
+    );
+  }
+
+  return (
+    <AnchorBase
+      variant={variant}
+      ref={ref as RefObject<HTMLAnchorElement>}
+      {...rest}
+    />
+  );
+});