+++ packages/gamut/src/Menu/elements.tsx
@@ -1,14 +1,19 @@
 import {
   styledOptions,
   system,
-  theme,
   transitionConcat,
 } from '@codecademy/gamut-styles';
 import { StyleProps, variance } from '@codecademy/variance';
 import styled from '@emotion/styled';
+import { ComponentProps, forwardRef } from 'react';
 
 import { sharedStates } from '../Box/props';
 import { resetStyles, Selectors } from '../ButtonBase/ButtonBase';
+
+enum MenuItemSelectors {
+  OUTLINE = '&:after',
+  OUTLINE_FOCUS_VISIBLE = '&:focus-visible:after',
+}
 
 type ListStyleProps = StyleProps<typeof listProps>;
 
@@ -34,7 +39,7 @@ export interface ListProps extends ListStyleProps, StyleStateProps {
   as?: 'ul' | 'ol';
 }
 
-export const List = styled('ul', styledOptions<'ul'>())<ListProps>(
+const StyledList = styled('ul', styledOptions<'ul'>())<ListProps>(
   system.css({
     listStyle: 'none',
     width: 1,
@@ -57,11 +62,12 @@ export const List = styled('ul', styledOptions<'ul'>())<ListProps>(
   listProps
 );
 
-List.defaultProps = {
-  root: true,
-  context: true,
-  m: 0,
-};
+export const List = forwardRef<
+  HTMLUListElement,
+  ComponentProps<typeof StyledList>
+>(({ context = true, m = 0, root = true, ...rest }, ref) => (
+  <StyledList context={context} m={m} root={root} ref={ref} {...rest} />
+));
 
 const interactiveVariants = system.variant({
   base: {
@@ -79,8 +85,20 @@ const interactiveVariants = system.variant({
       'linear'
     ),
     [Selectors.FOCUS_VISIBLE]: {
-      outline: `${theme.colors.primary} auto 2px`,
-      outlineOffset: '-2px',
+      outline: 'none',
+    },
+    [MenuItemSelectors.OUTLINE]: {
+      content: "''",
+      position: 'absolute',
+      inset: 4,
+      borderRadius: '4px',
+      border: 2,
+      borderColor: 'primary',
+      opacity: 0,
+      zIndex: 0,
+    },
+    [MenuItemSelectors.OUTLINE_FOCUS_VISIBLE]: {
+      opacity: 1,
     },
   },
   variants: {
@@ -166,7 +184,7 @@ export interface ListLinkProps extends ListItemProps {
   navlink?: boolean;
 }
 
-export const ListLink = styled('a', styledOptions<'a'>())<ListLinkProps>(
+const StyledListLink = styled('a', styledOptions<'a'>())<ListLinkProps>(
   resetStyles,
   interactiveVariants,
   activeStates,
@@ -175,9 +193,12 @@ export const ListLink = styled('a', styledOptions<'a'>())<ListLinkProps>(
   listProps
 );
 
-ListLink.defaultProps = {
-  zIndex: 1,
-};
+export const ListLink = forwardRef<
+  HTMLAnchorElement,
+  ComponentProps<typeof StyledListLink>
+>(({ zIndex = 1, ...rest }, ref) => (
+  <StyledListLink zIndex={zIndex} ref={ref} {...rest} />
+));
 
 export const ListButton = styled(
   'button',